Renames smart to legacy, standard to advanced
